requirements.txt
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
elasticsearch==7.10.0
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-logging==2.0.1
google-cloud-storage==1.33.0
google-cloud-vision==2.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.2
libcst==0.3.14
MarkupSafe==1.1.1
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-magic==0.4.18
python-snappy==0.5.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
